Adds article detail page - Issue #56
I extracted components for `ArticleContent` (all the stuff that was in the article card) and `ArticleActions` (the Edit, Link and Delete buttons) to allow the `ArticlePage` component to share as much of the existing layout as possible. The `ArticleContent` component has an `is_article_page` flag. If this is set to `true` then the article content starts off fully expanded, the abstract isn't truncated and the 'Show More' button is hidden.

There is still a bit of duplication between `ArticleList` and `ArticlePage` when it comes to the gallery and `ArticleEditor` logic. This is because the list shares one gallery and editor instance between all articles whereas the `ArticlePage` needs one for the individual article. This could possibly be improved later.