Add search icon button in search box - #55
Clicking on the search button takes the user to the search results page. I've added a new result at the top of the results list ('View all search results for *query*'), this is a workaround because the library used here (`react-select`) doesn't allow you to disable autofocussing yet - https://github.com/JedWatson/react-select/pull/3484

Making clicking on the search button work required another workaround that involves storing the query when the input is blurred (due to issues discussed here - https://github.com/JedWatson/react-select/issues/805). This means that clicking on the search button should work as expected but clicking outside the input and then later clicking the search button will go to the search page for the last stored query even though the input is empty. This isn't great but hopefully a rare enough case (why click on the search button for an empty search box?) that it's worth it to have it working like this for the moment.