Add a short `leaveDelay` to the MouseOverPopover to improve the situation where the popover would close when expanding truncated text
I think the issue here is that when the text expands the shape of the popup changes to accommodate it and this can mean the cursor is (mayber briefly) outside the popup so it would close. Adding the delay helps a bit but if the text expands so much that the tooltip has to move (e.g. render below the target because there's more space) then the tooltip will still close unexpectedly.