Add basic embedded viewer for HTML and PDF link types
There's no filtering in place yet to check for valid links.

The viewer and buttons to show it are only renderd on screens wider than the `lg` breakpoint (1920px).

The ArticleList and ArticleActions had to be changed to be responsive so that they would be rendered correctly when the viewer was visible.

I added the `unstated` library to allow sharing of the URL currently being viewed.