Move the ArticleActions to a dropdown menu on the Article card itself
This should fix the issue with the media viewer on mobile screens as well.

I removed the `ArticleLI` component because it's not necessary anymore now that the actions are handled directly in `ArticleContent`. This simplifies some of the component nesting that's going on when rendering articles but there's still more cleaning up to do on that front.