Fix bug where article details (video URLs etc.) weren't loading
The previous logic checked if the images were loaded and if they were assumed everything had been loaded. However, we load the images immediately now so they can be shown in the MiniFigureViewer so this check was failing. I've updated the logic so it's tracked in an explicit state variable (`details_fetched`) and this is set to `true` once a successful request has been made for the complete article data.